requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5 
autogen-agentchat>=0.1.10
colorama>=0.4.6